requirements.txt
transformers
tensorflow
pytorch
PIL